refactor(tests): refactorizar notification.processor tests (566 líneas → 4 archivos)
- notification.processor.spec.ts (566 líneas) → 4 archivos especializados:
  * notification.processor.order-confirmations.spec.ts (110 líneas)
  * notification.processor.payment-shipping.spec.ts (130 líneas)
  * notification.processor.welcome-priority.spec.ts (120 líneas)
  * notification.processor.errors-concurrency.spec.ts (160 líneas)
- Creado helper: helpers/notification-processor.test-helpers.ts (60 líneas)
- Tests: 19 passing
- Validaciones: lint ✅, type-check ✅, tests ✅, build ✅
